+++ docs/CHANGELOG.md
@@ -10,6 +10,10 @@ nav_order: 6
 
 ## main
 
+* Add `protocol` parameter to `with_request_url` test helper to enable testing with HTTPS protocol.
+
+    *Joel Hawksley*
+
 ## 4.3.0
 
 * Fix load order issues for 3rd-party template handlers.